+++ NEWS.md
@@ -21,6 +21,7 @@ The recoding and transformation functions get scoped variants, allowing to selec
 
 * `frq()` now removes empty columns before computing frequencies, because applying `frq()` on empty vectors caused an error.
 * `empty_cols()` and `empty_rows()` (and hence, `remove_empty_cols()` and `remove_empty_rows()`) caused an error for data frames with only one column resp. row, or if `x` was a vector and no data frame.
+* `frq()` now removes missing values from input when weights are applied, to ensure that input and weights have same length.
 
 
 # sjmisc 2.7.0